requirements.txt
flask
openai
PyMuPDF
python-dotenv
